replace %zu with %lu and btw fix some compile warnings
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@2149 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ cuda/sample_consensus/src/ransac.cpp
@@ -145,7 +145,7 @@ namespace pcl
       }
 
       if (debug_verbosity_level > 0)
-        fprintf (stderr, "[pcl::cuda::RandomSampleConsensus::computeModel] Model: %zu size, %d inliers.\n", model_.size (), n_best_inliers_count);
+        fprintf (stderr, "[pcl::cuda::RandomSampleConsensus::computeModel] Model: %lu size, %d inliers.\n", (unsigned long) model_.size (), n_best_inliers_count);
 
     //  if (model_.empty ())
     //  {
